requirements.txt
fastapi==0.54.2
mysql-connector-python==8.0.20
uvicorn==0.11.5
sqlalchemy==1.3.16
requests==2.23.0
python-consul==1.1.0
netifaces==0.10.9
lorem==0.1.1
pyjwt==1.7.1